Implement replay progress updates from upstreams
A node will now send periodic replay progress updates with the state
of all its slots. Currently it's done via WAL messaging, but we can
change that later if required.

During join we request more frequent updates via the join-target
node.

We need this during join so we can confirm we're safely caught up to idle peers
and their slots will be usable.  Without this we'll fail to join when any node
other than the join target is idle, it must do actual row-writing transactions
before we'll advance our replication origin. See comments in commit for
details.

We'll also use the same information later when we're doing node-loss
recovery and safe node part.