raft: Get rid of 2pc consenus and replace it with RAFT
I could not make the 2pc consensus work after couple of weeks so I gave
up and looked into RAFT. It seems that RAFT is actually easier to
implement than our original naive 2pc consensus idea.

This commit adds initial RAFT support and changes join to work with it.

There are a lot of cleanups pending and still several areas of
improvement but the current state is enough for tests to pass.

From the areas where improvements are worth mentioning:
 - we need to handle correctly the consensus worker exits/restarts, etc
 - the join code could be cleaned up, but I don't know it well enough
   to do it yet
 - the raft code is seriously underdocumented and is missing debug
   logging
 - consensus readme/general doc is now wrong in places where it still
   talks about 2pc consensus