bdr was failing to close opened relations when skipping changes
The BDR output plugin in the walsender would open the heap relation once for
each change and not close it again if the change was to be skipped. This leak
led to excessively slow processing of skipped tuples.  Profiling showed
excessive time spent in

    ResourceArrayAdd
    RelationIdGetRelation
    relation_open
    heap_open
    bdr_heap_open

and sample backtraces often look like

Per investigation done for RT#61404

Author: Craig Ringer.
Backpatching 6be6a66 from BDR2