Add MemCxt arg to bdr_nodes_get_local_info
... and have it allocate its output there, instead of changing to the
context ahead of time.  This allows it to use the current memory context
for transient stuff, and only switch to its output context when about
to create the permanent object.

The permanent memory context was being set to TopMemoryContext, and
apparently syscache accesses and catalog scans can leak tons of stuff
there that we don't want.