Return to previous memcxt after CommitTransactionCommand
A few places call StartTransactionCommand and end with
CommitTransactionCommand, which leaves the system pointing at
TopMemoryContext as the current context, which is not desirable because
any code that follows and was expecting to find a short-lived context is
now leaking forever.

We were already doing this in a few places; and perhaps some or all the
ones I'm patching don't need it, but it's hard to be sure.